fix: currenty-playing
