+++ src/commands/index.ts
@@ -1,8 +1,8 @@
 export * from "./help";
+export * from "./currently-playing";
 export * from "./info";
 export * from "./play";
 export * from "./pause-resume";
 export * from "./skip";
 export * from "./queue";
 export * from "./stop";
-export * from "./currentlyplaying";
