+++ src/commands/stop.ts
@@ -1,6 +1,7 @@
-import { CommandInteraction, GuildMember } from 'discord.js';
+import { CommandInteraction, EmbedBuilder, GuildMember } from 'discord.js';
 import { getVoiceConnection } from '@discordjs/voice';
 import { player } from '..';
+import { colors } from '../types';
 
 export const stopCommand = async (interaction: CommandInteraction) => {
     if (interaction.commandName === 'stop') {
@@ -8,12 +9,18 @@ export const stopCommand = async (interaction: CommandInteraction) => {
         const voiceChannel = member.voice.channel;
 
         if (!voiceChannel) {
-            return interaction.reply({ content: 'You need to be in a voice channel to stop the music.', ephemeral: true });
+            const embed = new EmbedBuilder()
+                .setColor(colors.Error)
+                .setDescription('You need to be in a voice channel to stop the music!');
+            return interaction.reply({ embeds: [embed], ephemeral: true });
         }
 
         const node = player.nodes.get(interaction.guildId!);
         if (!node) {
-            return interaction.reply({ content: 'No music is currently playing in this server.', ephemeral: true });
+            const embed = new EmbedBuilder()
+                .setColor(colors.Error)
+                .setDescription('No music is currently playing in this server.');
+            return interaction.reply({ embeds: [embed], ephemeral: true });
         }
 
         node.node.stop();
@@ -21,6 +28,9 @@ export const stopCommand = async (interaction: CommandInteraction) => {
         const connection = getVoiceConnection(interaction.guildId!);
         connection?.destroy();
 
-        await interaction.reply({ content: 'Stopped the music and left the voice channel.', ephemeral: true });
+        const embed = new EmbedBuilder()
+            .setColor(colors.Muusik)
+            .setDescription('Stopped the music and left the voice channel.');
+        await interaction.reply({ embeds: [embed], ephemeral: true });
     }
 };
